backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
httpx
python-dotenv
psycopg2-binary
python-multipart
